requirements.txt
mysql_connector_repackaged
python-dotenv
